🛠️ Added interceptors and filters

+++ src/users/users.service.ts
@@ -1,26 +1,23 @@
-import {
-  Injectable,
-  NotFoundException,
-  UnauthorizedException,
-} from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserRepository } from './repositories/user.repository';
 import { UserEntity } from './entities/user.entity';
+import { NotFoundError } from '../common/errors/types/NotFoundError';
+import { UnauthorizedError } from '../common/errors/types/UnauthorizedError';
 
 @Injectable()
 export class UsersService {
   constructor(private readonly repository: UserRepository) {}
 
   async create(createUserDto: CreateUserDto) {
-    // throw new UnauthorizedException('Não autorizado');
     return this.repository.create(createUserDto);
   }
 
   async findAll(): Promise<UserEntity[]> {
     const users = await this.repository.findAll();
     if (users.length === 0) {
-      throw new NotFoundException(
+      throw new NotFoundError(
         'Não foram localizados registros para o recurso solicitado',
       );
     }
@@ -31,7 +28,7 @@ export class UsersService {
   async findOne(id: string): Promise<UserEntity> {
     const user = await this.repository.findOne(id);
     if (!user) {
-      throw new NotFoundException('Usuário não encontrado na base de dados');
+      throw new NotFoundError('Usuário não encontrado na base de dados');
     }
 
     return user;
@@ -42,6 +39,11 @@ export class UsersService {
   }
 
   async remove(id: string) {
+    const user = await this.repository.findOne(id);
+    if (!user) {
+      throw new NotFoundError('Usuário não encontrado na base de dados');
+    }
+
     return this.repository.delete(id);
   }
 }
